Fix warnings: ../src/libslic3r/Emboss.cpp:858:13: warning: ‘Slic3r::EmbossStyle {anonymous}::create_style(const wstring&, const wstring&)’ defined but not used [-Wunused-function] ../src/libslic3r/Emboss.cpp:708:6: warning: ‘void {anonymous}::remove_small_islands(Slic3r::ExPolygons&, double)’ defined but not used [-Wunused-function] ../src/libslic3r/Emboss.cpp:497:6: warning: ‘void {anonymous}::visualize_heal(const string&, const ExPolygons&)’ defined but not used [-Wunused-function] ../src/libslic3r/Emboss.cpp:283:8: warning: ‘Slic3r::Points {anonymous}::collect_close_points(const ExPolygons&, double)’ defined but not used [-Wunused-function]

+++ src/slic3r/GUI/Gizmos/GLGizmoSVG.cpp
@@ -764,7 +764,7 @@ void draw_side_outline(const ExPolygons &shape, const std::array<unsigned char, 
 
     // overlap color
     auto draw = [&data, data_width, count_lines, get_offset, &color](int x, int y, float brightess) {
-        if (x < 0 || y < 0 || x >= data_width || y >= count_lines)
+        if (x < 0 || y < 0 || static_cast<size_t>(x) >= data_width || y >= count_lines)
             return; // out of image
         size_t offset = get_offset(x, y);
         bool change_color = false;
